# src/viewapp_v1/modules/config/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "REST API",
    "description": ""
  },
  "host": "localhost:3095",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
    "/": {
      "get": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/testDB": {
      "get": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/upload/{deviceID}/data": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "deviceID",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "hum",
            "in": "query",
            "type": "string"
          },
          {
            "name": "temp",
            "in": "query",
            "type": "string"
          },
          {
            "name": "tvoc",
            "in": "query",
            "type": "string"
          },
          {
            "name": "co",
            "in": "query",
            "type": "string"
          },
          {
            "name": "co2",
            "in": "query",
            "type": "string"
          },
          {
            "name": "pm25",
            "in": "query",
            "type": "string"
          },
          {
            "name": "o3",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/StatusGet/{deviceID}/powerStatus": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "deviceID",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/read/{deviceID}/ALL": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "deviceID",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/read/{deviceID}/hum": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "deviceID",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/read/{deviceID}/temp": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "deviceID",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/read/{deviceID}/tvoc": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "deviceID",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/read/{deviceID}/co2": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "deviceID",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/read/{deviceID}/co": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "deviceID",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/read/{deviceID}/o3": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "deviceID",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/switchCtr/{deviceID}/fan1": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "deviceID",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "status",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/switchCtr/{deviceID}/fan2": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "deviceID",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "status",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/statusRec/{deviceID}/viewALL": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "deviceID",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/statusNow/{deviceID}/viewfan1": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "deviceID",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/statusNow/{deviceID}/viewfan2": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "deviceID",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/statusNow/{deviceID}/viewALL": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "deviceID",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/CreateUser": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "username": {
                  "example": "any"
                },
                "password": {
                  "example": "any"
                },
                "LoginName": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/UpdateUserData": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "username": {
                  "example": "any"
                },
                "password": {
                  "example": "any"
                },
                "LoginName": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/Login": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "username": {
                  "example": "any"
                },
                "password": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  }
}